+++ sync-for-expenses/CodatSyncExpenses/Models/Shared/AccountStatus.cs
@@ -60,4 +60,5 @@ namespace Codat.Sync.Expenses.Models.Shared
             throw new Exception($"Unknown value {value} for enum AccountStatus");
         }
     }
+
 }