+++ sync-for-expenses/CodatSyncExpenses/Models/Shared/CompanyRequestBody.cs
@@ -11,6 +11,7 @@
 namespace Codat.Sync.Expenses.Models.Shared
 {
     using Newtonsoft.Json;
+    using System.Collections.Generic;
     
     public class CompanyRequestBody
     {
@@ -20,6 +21,12 @@ namespace Codat.Sync.Expenses.Models.Shared
         /// </summary>
         [JsonProperty("description")]
         public string? Description { get; set; }
+
+        /// <summary>
+        /// Reference to the groups that the company is assigned to.
+        /// </summary>
+        [JsonProperty("groups")]
+        public List<GroupItems>? Groups { get; set; }
 
         /// <summary>
         /// Name of company being connected.