+++ sync-for-expenses/CodatSyncExpenses/Models/Shared/TaxRateMappingInfoValidTransactionTypes.cs
@@ -65,4 +65,5 @@ namespace Codat.Sync.Expenses.Models.Shared
             throw new Exception($"Unknown value {value} for enum TaxRateMappingInfoValidTransactionTypes");
         }
     }
+
 }