+++ sync-for-expenses/CodatSyncExpenses/Models/Shared/TrackingRefDataType.cs
@@ -56,4 +56,5 @@ namespace Codat.Sync.Expenses.Models.Shared
             throw new Exception($"Unknown value {value} for enum TrackingRefDataType");
         }
     }
+
 }